UI updates
Editing Nav Code Structure to be more dynamic
