+++ src/components/NavLinks/NavLinks.scss
@@ -1,0 +1,2 @@
+@import '../../css/variables.scss';
+
